Debounce React.render
Prevents the reactComponent and reactDirective from calling React.render
multiple times before the function has a chance to complete.

For example, if the reactDirective invokes a render, and then before
that render is complete a second call to render is invoked, the timeout
will be cancelled and the new render call will begin.