Skip render on watcher init
Each watcher that is registered with the scope will invoke the listener
asynchronously to initialize the watcher. This means that the listener
may initially be called even through the value being watch has not
changed.

As a result, the render function may be called multiple times
unnecessarily. This change checks the case when the listener is invoked
during watcher initialization and prevents render from being called.

See PR davidchang/ngReact#56 for a discussion on this change.