Re-compute prop names on every compile
When ngReact compiles a directive, it computes a list of prop names. The
problem is that it assumes this list of prop names never changes: it is
computed once, when the directive is compiled for the first time.

As an example, let's consider this piece of HTML based on the
[HelloNoPropTypes](https://github.com/ngReact/ngReact/blob/b9475bd62def21d9b8662cc44e4f5880869d796d/tests/react-directive-spec.js#L77-L90)
test component:

```html
<hello-no-prop-types first-name="'Kent'"/>
<hello-no-prop-types first-name="'Bruce'" last-name="'Banner'"/>
```

The expected output is:

```
Hello Kent
Hello Bruce Banner
```

But the actual output is:

```
Hello Kent
Hello Bruce
```

This is beause on the first compile, the list of prop names ngReact
computes is `['firstName']`. When compiling the directive for the second
time, ngReact strips out the `lastName` prop because it doesn't belong
to the list of prop names computed in the first compile.

The fix this changeset provides is to re-compute the list of prop names
on every compile.

---

*Note 1*

This bug doesn't affect React components which expose propTypes. While
one could suggest to always set propTypes, this can be impractical, for
instance, when wrapping components with higher order components.

---

*Note 2*

A possible workaround for this bug is to manually provide the list of
prop names the component will ever receive to ngReact:

```js
reactDirective(HelloNoPropTypes, ['firstName', 'lastName'])
```

This is a bit cumbersome though.